% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lipcancer_offset}
\alias{lipcancer_offset}
\title{The vector of priori information for lip cancer case study}
\format{
\subsection{\code{lipcancer_offset}}{

A vector
}
}
\usage{
lipcancer_offset
}
\description{
The expected number of lip cancer cases in each Scotland county between 1975 and 1980.
}
\keyword{datasets}
