% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lipcancer_X}
\alias{lipcancer_X}
\title{The matrix of candidate covariates}
\format{
\subsection{\code{lipcancer_X}}{

A matrix with 56 observations and 1 covariate.
}
}
\usage{
lipcancer_X
}
\description{
This is a matrix with one candidate covariates, which is the proportion of population engaged in agriculture, fishing, or forestry.
}
\keyword{datasets}
