% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lipcancer_Y}
\alias{lipcancer_Y}
\title{The vector of observations for lip cancer case study}
\format{
\subsection{\code{lipcancer_Y}}{

A vector with 56 observations.
}
}
\usage{
lipcancer_Y
}
\description{
The observations about the observed number of lip cancer cases in each Scotland county between 1975 and 1980.
}
\keyword{datasets}
