% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lipcancer_Z}
\alias{lipcancer_Z}
\title{A list of design matrices}
\format{
\subsection{\code{lipcancer_Z}}{

A list of design matrices for random effects.
}
}
\usage{
lipcancer_Z
}
\description{
This is a list of two design matrices for two types of random effects.
The first component is for spatial random effects. The second componet is for overdispersion random effects.
}
\keyword{datasets}
