% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sigma}
\alias{Sigma}
\title{A list of covariance matrices}
\format{
\subsection{\code{Sigma}}{

A list of covariance matrices for random effects.
}
}
\usage{
Sigma
}
\description{
This is a list of two covariance matrices for two types of random effects.
The first component is for spatial random effects. The second componet is for overdispersion random effects.
}
\keyword{datasets}
