% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{X}
\alias{X}
\title{The matrix of candidate covariates}
\format{
\subsection{\code{X}}{

A matrix with 100 observations and 4 covariates.
}
}
\usage{
X
}
\description{
This is a matrix with four candidate covariates. The covariates in the first two columns are in the true model.
}
\keyword{datasets}
