% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Y}
\alias{Y}
\title{The vector of observations}
\format{
\subsection{\code{Y}}{

A vector with 100 observations.
}
}
\usage{
Y
}
\description{
The data was simulated under a Poisson Generalized linear mixed model with four candidate covariates, a vector of spatial random effects, and a vector of overdispersion random effects.
The matrix of candidate covariates, the design matrices for two types of random effects, and the covariance matrices for two types of random effects are provide in the package.
}
\keyword{datasets}
