% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Z}
\alias{Z}
\title{A list of design matrices}
\format{
\subsection{\code{Z}}{

A list of design matrices for random effects.
}
}
\usage{
Z
}
\description{
This is a list of two design matrices for two types of random effects.
The first component is for spatial random effects. The second componet is for overdispersion random effects.
}
\keyword{datasets}
